src/main/webapp/app/entities/admin-pages/route/admin-pages-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IAdminPages, AdminPages } from '../admin-pages.model';
import { AdminPagesService } from '../service/admin-pages.service';

import { AdminPagesRoutingResolveService } from './admin-pages-routing-resolve.service';

describe('AdminPages routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: AdminPagesRoutingResolveService;
  let service: AdminPagesService;
  let resultAdminPages: IAdminPages | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(AdminPagesRoutingResolveService);
    service = TestBed.inject(AdminPagesService);
    resultAdminPages = undefined;
  });

  describe('resolve', () => {
    it('should return IAdminPages returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAdminPages = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultAdminPages).toEqual({ id: 123 });
    });

    it('should return new IAdminPages if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAdminPages = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultAdminPages).toEqual(new AdminPages());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as AdminPages })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAdminPages = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultAdminPages).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
